Refactor - move resolver.utils.ts where it belongs
